TASK 4/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
streamlit-drawable-canvas>=0.9.0
plotly>=5.17.0
pyarrow>=16.1.0


